@@ -6,25 +6,26 @@ import com.google.gson.JsonSyntaxException;
 import mekanism.api.JsonConstants;
 import mekanism.api.SerializerHelper;
 import mekanism.api.chemical.gas.GasStack;
-import mekanism.api.recipes.inputs.chemical.GasStackIngredient;
+import mekanism.api.recipes.ingredients.ChemicalStackIngredient;
+import mekanism.api.recipes.ingredients.creator.IngredientCreatorAccess;
 import mekanism.common.Mekanism;
 import ml.northwestwind.fissionrecipe.misc.Heat;
 import ml.northwestwind.fissionrecipe.recipe.FissionRecipe;
-import net.minecraft.item.crafting.IRecipeSerializer;
-import net.minecraft.network.PacketBuffer;
-import net.minecraft.util.JSONUtils;
-import net.minecraft.util.ResourceLocation;
+import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.util.GsonHelper;
+import net.minecraft.world.item.crafting.RecipeSerializer;
 import net.minecraftforge.registries.ForgeRegistryEntry;
 
 import javax.annotation.Nullable;
 import javax.script.ScriptException;
 
-public class FissionRecipeSerializer extends ForgeRegistryEntry<IRecipeSerializer<?>> implements IRecipeSerializer<FissionRecipe> {
+public class FissionRecipeSerializer extends ForgeRegistryEntry<RecipeSerializer<?>> implements RecipeSerializer<FissionRecipe> {
     @Override
     public FissionRecipe fromJson(ResourceLocation recipeId, JsonObject json) {
-        JsonElement input = JSONUtils.isArrayNode(json, JsonConstants.INPUT) ? JSONUtils.getAsJsonArray(json, JsonConstants.INPUT) :
-                JSONUtils.getAsJsonObject(json, JsonConstants.INPUT);
-        GasStackIngredient inputIngredient = GasStackIngredient.deserialize(input);
+        JsonElement input = GsonHelper.isArrayNode(json, JsonConstants.INPUT) ? GsonHelper.getAsJsonArray(json, JsonConstants.INPUT) :
+                GsonHelper.getAsJsonObject(json, JsonConstants.INPUT);
+        ChemicalStackIngredient.GasStackIngredient inputIngredient = IngredientCreatorAccess.gas().deserialize(input);
         GasStack output = SerializerHelper.getGasStack(json, JsonConstants.OUTPUT);
         JsonElement heatObj = json.get("heat");
         boolean isEqt = false;
@@ -49,9 +50,9 @@ public class FissionRecipeSerializer extends ForgeRegistryEntry<IRecipeSerialize
 
     @Nullable
     @Override
-    public FissionRecipe fromNetwork(ResourceLocation recipeId, PacketBuffer buffer) {
+    public FissionRecipe fromNetwork(ResourceLocation recipeId, FriendlyByteBuf buffer) {
         try {
-            GasStackIngredient inputIngredient = GasStackIngredient.read(buffer);
+            ChemicalStackIngredient.GasStackIngredient inputIngredient = IngredientCreatorAccess.gas().read(buffer);
             GasStack output = GasStack.readFromPacket(buffer);
             boolean isEqt = buffer.readBoolean();
             Heat heat = new Heat(isEqt, isEqt ? 0 : buffer.readFloat(), isEqt ? buffer.readUtf() : null);
@@ -63,7 +64,7 @@ public class FissionRecipeSerializer extends ForgeRegistryEntry<IRecipeSerialize
     }
 
     @Override
-    public void toNetwork(PacketBuffer buffer, FissionRecipe recipe) {
+    public void toNetwork(FriendlyByteBuf buffer, FissionRecipe recipe) {
         try {
             recipe.write(buffer);
         } catch (Exception e) {
